capmarket-data-app-main/workflows/requirements.txt
oauthlib
requests
requests-oauthlib
PyJWT
urllib3
sseclient-py==1.7